+++ doc/ReleaseNotes.md
@@ -1,6 +1,11 @@
 # SVG.NET Release Notes
 The release versions are NuGet releases.
 
+## [Version X.X.X](https://www.nuget.org/packages/Svg/X.X.X)  (20XX-XX-XX)
+
+### Fixes
+* fixed XML namespace prefixes are also applied for nodes declaring them (see [PR #1106](https://github.com/svg-net/SVG/pull/1106))
+
 ## [Version 3.4.6](https://www.nuget.org/packages/Svg/3.4.6)  (2023-11-16)
 
 ### Fixes